Previous behavior caused dc.GetMultiLineTextExtent to fail on Mac.
Previous behavior did not work on Mac because dc.GetMultiLineTextExtent
does not work there until an object is selected. The wx documentation states:
"...If a bitmap is not given to this constructor then don’t forget to select
a bitmap into the DC before drawing on it." so the Mac behavior is proper
even if this step can be skipped elsewhere without a problem. The fix is
simply to select an empty bitmap first.

Note: This commit was created by Stefan Hacker from a patch sent by the
author and is slightly modified. The author was set for proper attribution.